Add object status accessor interface.
